Update sql.py
Good milestone to check in.

+++ practice/sql.py
@@ -1,17 +1,63 @@
-"""TO do
+"""To Do  
 """
-
 
 import sqlite3
 
+#Helper Functions
+def select_all_from_table(tablename, sql_cursor):
+    """ Prints all results from a table named tablename.
+        Args: table_name - table name to query (Case sensative)
+              sql_cursor - sqlite3 cursor object
+        Returns: Doesn't return, but will print all results.
+    """
+    
+    sql = "SELECT * FROM {0}".format(tablename)
+    sql_cursor.execute(sql)
+    # print("fetchall:")
+    all_results = sql_cursor.fetchall()
+    for result in all_results:
+        print(result)
+
+def run_sql_command_return_all_results(command, sql_cursor):
+    """ Prints all results of 
+    """
+    sql_cursor.execute(command)
+    # print("fetchall:")
+    all_results = sql_cursor.fetchall()
+    for result in all_results:
+        print(result)
+
+def query_if_table_exists(table_name):
+    """ Query if a table exsists in the sqlite)master meta data table.
+        Args: table_name - table name to query (Case sensative)
+        Returns: Length of the list object that contains the tables that matched.
+    """
+
+    SQL_COMMAND="SELECT sql FROM sqlite_master WHERE type='table' AND name ='{0}'".format(table_name)
+    CURSOR.execute(SQL_COMMAND)
+    all_results = CURSOR.fetchall()
+    #print(all_results)
+    return(len(all_results))
+
+
+#Create a database and a connection object
+    
 CONNECTION = sqlite3.connect("sounds_like_a_gym.db")
 CURSOR = CONNECTION.cursor()
 
 
-CURSOR.execute("""DROP TABLE Members;""")
-CURSOR.execute("""DROP TABLE Organization;""")
+#If this file is rerun, the tables will need to be dropped if they exsist since
+#the database exsists in a file and not in memory
+# Running the database in memory has certain advantanges and disadvantages
 
+if query_if_table_exists("Members"):
+           CURSOR.execute("""DROP TABLE Members;""")
 
+if query_if_table_exists("Organization"):
+           CURSOR.execute("""DROP TABLE Organization;""")
+
+#Functions to create the Members and Organization tables
+           
 SQL_COMMAND = """
 CREATE TABLE Members (
 ID INTEGER PRIMARY KEY,
@@ -37,6 +83,10 @@ DUES DECIMAL(19,4));"""
 
 CURSOR.execute(SQL_COMMAND)
 
+
+
+
+#Sample Data Table used for testing
 MEMBERS_DATA = [("William", "Shakespeare",
                  "1000 Innovation Way",
                  "Apt 400", "Fake London",
@@ -73,6 +123,11 @@ MEMBERS_DATA = [("William", "Shakespeare",
                  "Georgia", "11009",
                  "111-222-333", "22"), ]
 
+#Load the sample data into the databse
+# I pictured this as a gym member who was signing up for a gym at a specific location.
+# The location is in Alpharetta and the dues are 30 if the member is under 65.
+# The dues are setup that way so that there would be members with 0 dues.
+
 for member in MEMBERS_DATA:
     members_format_str = (
         'INSERT INTO members(ID, FIRST_NAME,'
@@ -98,7 +153,7 @@ for member in MEMBERS_DATA:
     CURSOR.execute(SQL_COMMAND)
     result_insert = CURSOR.fetchone()
     members_table_id = result_insert[0]
-
+    
     gym_location = "Alpharetta"
     gym_dues = 30.00
 
@@ -114,30 +169,9 @@ for member in MEMBERS_DATA:
                                                  dues=gym_dues)
 
     CURSOR.execute(SQL_COMMAND)
-
-
-def select_all_from_table(tablename, sql_cursor):
-    """ Selects all
-    """
-    sql = "SELECT * FROM {0}".format(tablename)
-    sql_cursor.execute(sql)
-    # print("fetchall:")
-    all_results = sql_cursor.fetchall()
-    for result in all_results:
-        print(result)
-
-
-def run_sql_command_return_all_results(command, sql_cursor):
-    """ Selects all
-    """
-    sql_cursor.execute(command)
-    # print("fetchall:")
-    all_results = sql_cursor.fetchall()
-    for result in all_results:
-        print(result)
-
-
-#
+    
+# See if the tables are filled with the correct data from the sample
+# dataset
 select_all_from_table("Members", CURSOR)
 select_all_from_table("Organization", CURSOR)
